+++ lib/kitchen/directions/bake_notes/bake_numbered_notes/main.rb
@@ -11,8 +11,15 @@ module Kitchen
         V2.new.bake(book: book, classes: classes)
       end
 
-      # Used by V1, V2
-      def self.bake_note_exercise(note:, exercise:, divider: ' ')
+      # V3 bakes notes tied to an example immediately previous ("Try It" notes)
+      # Must be called AFTER BakeExercises
+      #
+      def self.v3(book:, classes:, suppress_solution: true)
+        V3.new.bake(book: book, classes: classes, suppress_solution: suppress_solution)
+      end
+
+      # Used by V1, V2, V3
+      def self.bake_note_exercise(note:, exercise:, divider: ' ', suppress_solution: false)
         exercise.add_class('unnumbered')
         # bake problem
         exercise.problem.wrap_children('div', class: 'os-problem-container')
@@ -20,8 +27,16 @@ module Kitchen
         return unless exercise.solution
 
         # bake solution in place
-        BakeNumberedExercise.bake_solution_v1(
-          exercise: exercise, number: note.first('.os-number').text.gsub(/#/, ''), divider: divider)
+        if suppress_solution
+          exercise.add_class('os-hasSolution')
+          exercise.solution.trash
+        else
+          BakeNumberedExercise.bake_solution_v1(
+            exercise: exercise,
+            number: note.first('.os-number').text.gsub(/#/, ''),
+            divider: divider
+          )
+        end
       end
     end
   end