+++ CHANGELOG.md
@@ -22,12 +22,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added spec for translations (minor)
 * Remove summary attribute from numbered tables, add option to pass title element to `BakePreface` (minor)
 * Renamed `American Government` strategy to `Default` inside `move_solutions_to_answer_key` for other books to use it with a sent in classname (major)
+* Adds `#previous` method to note_elements to find the immediately previous element (minor)
+* Adds `BakeNumberedNotes` V3 (minor)
 * Added line that puts the classname `has-footnote` in the footnote ref's parent element (major)
 * Added a condition into `BakeChapterSummary` so it doesn't bake the title if it already includes the respective number in it
 * Create v3 of MoveExercisesToEOC which differs from v1 by the presence of a section title
 and from v2 the lack of additional "os-section-area" and os-#{@klass} wrapper (minor)
 * Add a condition in BakeNumberedExercise to make it possible to suppress even solutions in the Answer Key (minor)
 * Fix BakeFurtherResearch baking with main bake script error by breaking the loop if further research sections are not present (minor)
+
 ## [5.0.0] - 2021-06-02
 
 * Reditributed duplicated id logic across `#record_id_copied`, `#record_id_cut`, `#record_id_pasted`, added a couple more tests for `#copy`, `#cut`, and `#paste`, created a new class `IdTracker` and moved `#record_id_copied`, `#record_id_cut`, `#record_id_pasted`, and `modified_id_to_paste` into the new class (major)